Allow to request certs without OCSP Must-Staple extension
This change adds a $ocsp_must_staple flag (defaults to true like before)
that allows to generate certificate requests that don't request the
OCSP Must-Staple extension.

This is useful for services that don't (yet) support OCSP stapling,
since compliant clients will otherwise reject the certificate if stapled
OCSP information is not provided by the service.

Furthermore, with this change, the keyUsage extension entries are now
always requested, even if no SAN are available.