fix possible conflicts in cert deployment, refs #24
When using this module to create certificates *for* the Puppet Master,
then it would try to deploy all signed certificates locally. This would
fail, causing the Puppet run on the Puppet Master to abort.

It should be noted that this specific conflict would only occur when
the certificate on the Puppet Master has the same name as $acme_host.